Remove Cursor from Screen completely if type is None
Previously chrome would load a 1x1 dummy image if
cursor was none, this has changed to now remove the cursor
from screen.
Since all the code passing the cursor around for that path
deals with HCURSOR, which in Microsoft Documentation is specificed
as  'If this parameter is NULL, the cursor is removed from the screen.'
this should be the correct thing to do here.